requirements.txt
aiogram==3.13.1
SQLAlchemy[asyncio]==2.0.36
aiomysql==0.2.0

alembic==1.14.0
python-dotenv==1.0.1
cryptography==44.0.0
